feat(errorBoundary): Separated error boundary component of hoc
